Modified

+++ core/update.py
@@ -153,7 +153,7 @@ class Decoder(nn.Module):
             # w_embeds = Parameter(torch.empty((1, hidden_dim // 2, 1, w))).repeat((1, 1, h, 1))
             # self.pos_embeds = torch.cat((h_embeds, w_embeds), dim=1)
             self.pos_embeds = nn.Embedding(num_embeddings=w * h, embedding_dim=hidden_dim)
-            self.pos_embeds = self.pos_embeds(torch.range(start=0, end=w * h))
+            self.pos_embeds = self.pos_embeds(torch.range(start=0, end=w * h, dtype=torch.int))
             self.pos_embeds.view((1, h, w, hidden_dim))
 
     def forward(self, query, key):
